refactor: editor_types.ts를 barrel로 교체 (#266)
